initial pass of refactor for making the crate extensible over multiple option valuation models
