+++ Cargo.toml
@@ -1,19 +1,22 @@
+########## Crate Info ##########
 [package]
 name = "blackscholes"
 version = "0.24.0"
 edition = "2021"
 license = "MIT"
-description = "Black-Scholes option pricing model calculator"
+description = "Option valuation models"
 repository = "https://github.com/hayden4r4/blackscholes-rust"
 documentation = "https://docs.rs/blackscholes"
 readme = "README.md"
 authors = ["Hayden Rose"]
 keywords = ["finance", "option", "pricing", "blackscholes", "option-pricing"]
 
+########## Compilation Configuration ##########
 [profile.release]
 codegen-units = 1
 lto = "fat"
 
+########## Dependencies ##########
 [dev-dependencies]
 approx = "0.5.1"
 assert_approx_eq = "1.1.0"
@@ -29,11 +32,8 @@ statrs = "0.17.1"
 getrandom = { version = "0.2", features = ["js"] }
 
 
-#[[bench]]
-#name = "black"
-#harness = false
 
-# Benchmarks
+########## Benchmarks ##########
 [[bench]]
 name = "pricing"
 harness = false
@@ -41,6 +41,10 @@ harness = false
 [[bench]]
 name = "implied_volatility"
 harness = false
+
+#[[bench]]
+#name = "black"
+#harness = false
 
 # enable this to run additional benches
 # [[bench]]
