+++ src/constants.rs
@@ -1,0 +1,4 @@
+//! Global constants
+pub use statrs::consts::SQRT_2PI;
+pub use std::f64::consts::PI;
+pub(crate) const DAYS_PER_YEAR: f64 = 365.25;
