+++ src/lets_be_rational/black.rs
@@ -2,15 +2,13 @@ use std::f64::consts::FRAC_1_SQRT_2;
 
 use statrs::consts::SQRT_2PI;
 
-use crate::{
-    lets_be_rational::{
-        cody::optimized::{erfc, erfcx},
-        intrinsic::normalised_intrinsic,
-        normal_distribution::standard_normal_cdf,
-        DENORMALISATION_CUTOFF, ONE_OVER_SQRT_TWO_PI,
-    },
-    OptionType,
+use crate::lets_be_rational::{
+    cody::optimized::{erfc, erfcx},
+    intrinsic::normalised_intrinsic,
+    normal_distribution::standard_normal_cdf,
+    DENORMALISATION_CUTOFF, ONE_OVER_SQRT_TWO_PI,
 };
+use crate::OptionType;
 
 const H_LARGE: f64 = -10.0;
