+++ src/valuators/black_scholes/constants.rs
@@ -1,0 +1,7 @@
+//! Local Constants
+pub(crate) const A: f64 = 4.626_275_3e-1;
+pub(crate) const B: f64 = -1.168_519_2e-2;
+pub(crate) const C: f64 = 9.635_418_5e-4;
+pub(crate) const D: f64 = 7.535_022_5e-5;
+pub(crate) const _E: f64 = 1.424_516_45e-5;
+pub(crate) const F: f64 = -2.102_376_9e-5;
