+++ src/valuators/black_scholes/greeks.rs
@@ -2,8 +2,9 @@ use std::collections::HashMap;
 
 use num_traits::Float;
 
-use crate::{Inputs, OptionType, Pricing, *};
-
+use crate::constants::*;
+use crate::valuators::black_scholes::{distributions::*, Inputs, Pricing};
+use crate::OptionType;
 pub trait Greeks<T>: Pricing<T>
 where
     T: Float,
@@ -36,7 +37,8 @@ impl Greeks<f64> for Inputs {
     /// f64 of the delta of the option.
     /// # Example
     /// ```
-    /// use blackscholes::{Inputs, OptionType, Greeks};
+    /// use blackscholes::OptionType;
+    /// use blackscholes::valuators::black_scholes::{Inputs, Greeks};
     /// let inputs = Inputs::new(OptionType::Call, 100.0, 100.0, None, 0.05, 0.2, 20.0/365.25, Some(0.2));
     /// let delta = inputs.calc_delta().unwrap();
     /// ```
@@ -55,7 +57,8 @@ impl Greeks<f64> for Inputs {
     /// f64 of the gamma of the option.
     /// # Example
     /// ```
-    /// use blackscholes::{Inputs, OptionType, Greeks};
+    /// use blackscholes::OptionType;
+    /// use blackscholes::valuators::black_scholes::{Inputs, Greeks};
     /// let inputs = Inputs::new(OptionType::Call, 100.0, 100.0, None, 0.05, 0.2, 20.0/365.25, Some(0.2));
     /// let gamma = inputs.calc_gamma().unwrap();
     /// ```
@@ -77,7 +80,8 @@ impl Greeks<f64> for Inputs {
     /// f64 of theta per day (not per year).
     /// # Example
     /// ```
-    /// use blackscholes::{Inputs, OptionType, Greeks};
+    /// use blackscholes::OptionType;
+    /// use blackscholes::valuators::black_scholes::{Inputs, Greeks};
     /// let inputs = Inputs::new(OptionType::Call, 100.0, 100.0, None, 0.05, 0.2, 20.0/365.25, Some(0.2));
     /// let theta = inputs.calc_theta().unwrap();
     /// ```
@@ -106,7 +110,8 @@ impl Greeks<f64> for Inputs {
     /// f64 of the vega of the option.
     /// # Example
     /// ```
-    /// use blackscholes::{Inputs, OptionType, Greeks};
+    /// use blackscholes::OptionType;
+    /// use blackscholes::valuators::black_scholes::{Inputs, Greeks};
     /// let inputs = Inputs::new(OptionType::Call, 100.0, 100.0, None, 0.05, 0.2, 20.0/365.25, Some(0.2));
     /// let vega = inputs.calc_vega().unwrap();
     /// ```
@@ -123,7 +128,8 @@ impl Greeks<f64> for Inputs {
     /// f64 of the rho of the option.
     /// # Example
     /// ```
-    /// use blackscholes::{Inputs, OptionType, Greeks};
+    /// use blackscholes::OptionType;
+    /// use blackscholes::valuators::black_scholes::{Inputs, Greeks};
     /// let inputs = Inputs::new(OptionType::Call, 100.0, 100.0, None, 0.05, 0.2, 20.0/365.25, Some(0.2));
     /// let rho = inputs.calc_rho().unwrap();
     /// ```
@@ -148,7 +154,8 @@ impl Greeks<f64> for Inputs {
     /// f64 of the epsilon of the option.
     /// # Example
     /// ```
-    /// use blackscholes::{Inputs, OptionType, Greeks};
+    /// use blackscholes::OptionType;
+    /// use blackscholes::valuators::black_scholes::{Inputs, Greeks};
     /// let inputs = Inputs::new(OptionType::Call, 100.0, 100.0, None, 0.05, 0.2, 20.0/365.25, Some(0.2));
     /// let epsilon = inputs.calc_epsilon().unwrap();
     /// ```
@@ -168,7 +175,8 @@ impl Greeks<f64> for Inputs {
     /// f64 of the lambda of the option.
     /// # Example
     /// ```
-    /// use blackscholes::{Inputs, OptionType, Greeks, Pricing};
+    /// use blackscholes::OptionType;
+    /// use blackscholes::valuators::black_scholes::{Inputs, Greeks, Pricing};
     /// let inputs = Inputs::new(OptionType::Call, 100.0, 100.0, None, 0.05, 0.2, 20.0/365.25, Some(0.2));
     /// let lambda = inputs.calc_lambda().unwrap();
     /// ```
@@ -184,7 +192,8 @@ impl Greeks<f64> for Inputs {
     /// f64 of the vanna of the option.
     /// # Example
     /// ```
-    /// use blackscholes::{Inputs, OptionType, Greeks};
+    /// use blackscholes::OptionType;
+    /// use blackscholes::valuators::black_scholes::{Inputs, Greeks};
     /// let inputs = Inputs::new(OptionType::Call, 100.0, 100.0, None, 0.05, 0.2, 20.0/365.25, Some(0.2));
     /// let vanna = inputs.calc_vanna().unwrap();
     /// ```
@@ -206,7 +215,8 @@ impl Greeks<f64> for Inputs {
     // /// f64 of the charm of the option.
     // /// # Example
     /// ```
-    /// use blackscholes::{Inputs, OptionType, Greeks};
+    /// use blackscholes::OptionType;
+    /// use blackscholes::valuators::black_scholes::{Inputs, Greeks};
     /// let inputs = Inputs::new(OptionType::Call, 100.0, 100.0, None, 0.05, 0.2, 20.0/365.25, Some(0.2));
     /// let charm = inputs.calc_charm().unwrap();
     /// ```
@@ -233,7 +243,8 @@ impl Greeks<f64> for Inputs {
     /// f64 of the veta of the option.
     /// # Example
     /// ```
-    /// use blackscholes::{Inputs, OptionType, Greeks};
+    /// use blackscholes::OptionType;
+    /// use blackscholes::valuators::black_scholes::{Inputs, Greeks};
     /// let inputs = Inputs::new(OptionType::Call, 100.0, 100.0, None, 0.05, 0.2, 20.0/365.25, Some(0.2));
     /// let veta = inputs.calc_veta().unwrap();
     /// ```
@@ -261,7 +272,8 @@ impl Greeks<f64> for Inputs {
     /// f64 of the vomma of the option.
     /// # Example
     /// ```
-    /// use blackscholes::{Inputs, OptionType, Greeks};
+    /// use blackscholes::OptionType;
+    /// use blackscholes::valuators::black_scholes::{Inputs, Greeks};
     /// let inputs = Inputs::new(OptionType::Call, 100.0, 100.0, None, 0.05, 0.2, 20.0/365.25, Some(0.2));
     /// let vomma = inputs.calc_vomma().unwrap();
     /// ```
@@ -282,7 +294,8 @@ impl Greeks<f64> for Inputs {
     /// f64 of the speed of the option.
     /// # Example
     /// ```
-    /// use blackscholes::{Inputs, OptionType, Greeks};
+    /// use blackscholes::OptionType;
+    /// use blackscholes::valuators::black_scholes::{Inputs, Greeks};
     /// let inputs = Inputs::new(OptionType::Call, 100.0, 100.0, None, 0.05, 0.2, 20.0/365.25, Some(0.2));
     /// let speed = inputs.calc_speed().unwrap();
     /// ```
@@ -304,7 +317,8 @@ impl Greeks<f64> for Inputs {
     /// f64 of the zomma of the option.
     /// # Example
     /// ```
-    /// use blackscholes::{Inputs, OptionType, Greeks};
+    /// use blackscholes::OptionType;
+    /// use blackscholes::valuators::black_scholes::{Inputs, Greeks};
     /// let inputs = Inputs::new(OptionType::Call, 100.0, 100.0, None, 0.05, 0.2, 20.0/365.25, Some(0.2));
     /// let zomma = inputs.calc_zomma().unwrap();
     /// ```
@@ -326,7 +340,8 @@ impl Greeks<f64> for Inputs {
     /// f64 of the color of the option.
     /// # Example
     /// ```
-    /// use blackscholes::{Inputs, OptionType, Greeks};
+    /// use blackscholes::OptionType;
+    /// use blackscholes::valuators::black_scholes::{Inputs, Greeks};
     /// let inputs = Inputs::new(OptionType::Call, 100.0, 100.0, None, 0.05, 0.2, 20.0/365.25, Some(0.2));
     /// let color = inputs.calc_color().unwrap();
     /// ```
@@ -355,7 +370,8 @@ impl Greeks<f64> for Inputs {
     /// f64 of the ultima of the option.
     /// # Example
     /// ```
-    /// use blackscholes::{Inputs, OptionType, Greeks};
+    /// use blackscholes::OptionType;
+    /// use blackscholes::valuators::black_scholes::{Inputs, Greeks};
     /// let inputs = Inputs::new(OptionType::Call, 100.0, 100.0, None, 0.05, 0.2, 20.0/365.25, Some(0.2));
     /// let ultima = inputs.calc_ultima().unwrap();
     /// ```
@@ -377,7 +393,8 @@ impl Greeks<f64> for Inputs {
     /// f64 of the dual delta of the option.
     /// # Example
     /// ```
-    /// use blackscholes::{Inputs, OptionType, Greeks};
+    /// use blackscholes::OptionType;
+    /// use blackscholes::valuators::black_scholes::{Inputs, Greeks};
     /// let inputs = Inputs::new(OptionType::Call, 100.0, 100.0, None, 0.05, 0.2, 20.0/365.25, Some(0.2));
     /// let dual_delta = inputs.calc_dual_delta().unwrap();
     /// ```
@@ -399,7 +416,8 @@ impl Greeks<f64> for Inputs {
     /// f64 of the dual gamma of the option.
     /// # Example
     /// ```
-    /// use blackscholes::{Inputs, OptionType, Greeks};
+    /// use blackscholes::OptionType;
+    /// use blackscholes::valuators::black_scholes::{Inputs, Greeks};
     /// let inputs = Inputs::new(OptionType::Call, 100.0, 100.0, None, 0.05, 0.2, 20.0/365.25, Some(0.2));
     /// let dual_gamma = inputs.calc_dual_gamma().unwrap();
     /// ```
@@ -421,7 +439,8 @@ impl Greeks<f64> for Inputs {
     /// HashMap of type <String, f64> of all Greeks of the option.
     /// # Example
     /// ```
-    /// use blackscholes::{Inputs, OptionType, Greeks};
+    /// use blackscholes::OptionType;
+    /// use blackscholes::valuators::black_scholes::{Inputs, Greeks};
     /// let inputs = Inputs::new(OptionType::Call, 100.0, 100.0, None, 0.05, 0.2, 20.0/365.25, Some(0.2));
     /// let greeks = inputs.calc_all_greeks().unwrap();
     /// ```
